+++ club-api/src/test/java/com/smh/club/api/rest/mappers/EmailMapperTests.java
@@ -2,7 +2,9 @@ package com.smh.club.api.rest.mappers;
 
 import com.smh.club.api.rest.config.MapperConfig;
 import com.smh.club.api.rest.domain.entities.EmailEntity;
-import com.smh.club.api.rest.dto.EmailDto;
+import com.smh.club.api.rest.dto.email.EmailCreateDto;
+import com.smh.club.api.rest.dto.email.EmailDto;
+import com.smh.club.api.rest.dto.email.EmailUpdateDto;
 import org.instancio.Instancio;
 import org.instancio.junit.InstancioExtension;
 import org.instancio.junit.WithSettings;
@@ -33,7 +35,7 @@ public class EmailMapperTests {
     @Test
     public void from_createDto_to_entity() {
         // setup
-        var create = Instancio.of(EmailDto.class)
+        var create = Instancio.of(EmailCreateDto.class)
                 .ignore(field(EmailDto::getId))
                 .create();
 
@@ -87,9 +89,9 @@ public class EmailMapperTests {
     }
 
     @Test
-    public void update_entity_from_createDto() {
+    public void update_entity_from_updateDto() {
         // setup
-        var update = Instancio.create(EmailDto.class);
+        var update = Instancio.create(EmailUpdateDto.class);
         var entity = Instancio.create(EmailEntity.class);
 
         // execute
